save args have problems

+++ app/src/main/java/com/udacity/shoestore/Data/Shoe.kt
@@ -1,3 +1,3 @@
 package com.udacity.shoestore.Data
 
-class Shoe(var name: String, var size: Int, var Company: String, var Discription:String)
+class Shoe(var name: String, var size: Float, var Company: String, var Discription:String)
